refactor all atom types to use factory classes
This is quite a large refactoring of the MP4 parser.

It splits every atom into a class that represents the data in the atom
and a factory that can generate the atom from a BinaryIO source.

The Mp4Atom.load() function is moved to IsoParser and split into two
functions:

* IsoParser.load()
* IsoParser.load_wrapped()

The access to atoms inside an atom is changed from attribute access
to item access. For example wrap.moov becomes wrap["moov]". This
removes a large source of infinite recursion problems where access to
attributes of an Mp4Atom had the side-effect of trying to search for
children.

The encode() function now requires a destination BinaryIO. To produce
a byte array, there is a new encode_as_bytes() function.

+++ dashlive/server/models/mediafile.py
@@ -113,7 +113,7 @@ class MediaFile(ModelMixin["MediaFile"], Base):
                 self._representation = None
         return self._representation
 
-    def set_representation(self, rep):
+    def set_representation(self, rep: Representation) -> None:
         self.rep = rep.toJSON(pure=True)
         self._representation = rep
 
@@ -164,10 +164,14 @@ class MediaFile(ModelMixin["MediaFile"], Base):
                   size: int,
                   buffer_size: int = Blob.BUFFER_SIZE) -> contextlib.AbstractContextManager[BinaryIO]:
         abs_path = self.absolute_path(self.stream.directory)
+        if self.blob is None:
+            raise IOError('MediaFile has no blob')
         return self.blob.open_file(abs_path, start=start, size=size, buffer_size=buffer_size)
 
     def delete_file(self) -> None:
         abs_path = self.absolute_path(self.stream.directory)
+        if self.blob is None:
+            raise IOError('MediaFile has no blob')
         self.blob.delete_file(abs_path)
 
     def as_stream_timing_reference(self) -> StreamTimingReference | None:
@@ -321,11 +325,12 @@ class MediaFile(ModelMixin["MediaFile"], Base):
                         src.seek(frag.pos)
                         reader = BufferedReader(
                             src, offset=frag.pos, size=frag.size, buffersize=16384)
-                        atom = mp4.IsoParser.load(
-                            reader, options=mp4_options, use_wrapper=True)
+                        atom: mp4.Wrapper = mp4.IsoParser.load_wrapped(
+                            reader, options=mp4_options)
+                        reader.close()
                         changed = modify_atoms(atom)
                         if changed:
-                            dest.write(atom.encode())
+                            atom.encode(dest)
                         else:
                             src.seek(frag.pos)
                             dest.write(src.read(frag.size))
